Refactor providers to remove duplicate implementations
